Modularized preferences into fragments and individual xml files. Preferences are now Screens: Profile, Naming, Workflow, Printing, Database, BrAPI, and About which go to individual fragments.
