+++ app/src/main/java/org/phenoapps/intercross/fragments/preferences/ProfileFragment.kt
@@ -1,0 +1,6 @@
+package org.phenoapps.intercross.fragments.preferences
+
+import org.phenoapps.intercross.R
+
+class ProfileFragment : ToolbarPreferenceFragment(R.xml.profile_preferences,
+    "org.phenoapps.intercross.ROOT_PREFERENCES_PROFILE")
