+++ app/src/main/res/drawable/ic_account_tree_black_24dp.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M22,11V3h-7v3H9V3H2v8h7V8h2v10h4v3h7v-8h-7v3h-2V8h2v3H22zM7,9H4V5h3V9zM17,15h3v4h-3V15zM17,5h3v4h-3V5z"
+      android:fillColor="#000000"/>
+</vector>
