+++ app/src/main/res/drawable/ic_badge_black_24dp.xml
@@ -1,0 +1,21 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M14,12h4v1.5h-4z"
+      android:fillColor="#000000"/>
+  <path
+      android:pathData="M14,15h4v1.5h-4z"
+      android:fillColor="#000000"/>
+  <path
+      android:pathData="M20,7h-5V4c0,-1.1 -0.9,-2 -2,-2h-2C9.9,2 9,2.9 9,4v3H4C2.9,7 2,7.9 2,9v11c0,1.1 0.9,2 2,2h16c1.1,0 2,-0.9 2,-2V9C22,7.9 21.1,7 20,7zM11,7V4h2v3v2h-2V7zM20,20H4V9h5c0,1.1 0.9,2 2,2h2c1.1,0 2,-0.9 2,-2h5V20z"
+      android:fillColor="#000000"/>
+  <path
+      android:pathData="M9,13.5m-1.5,0a1.5,1.5 0,1 1,3 0a1.5,1.5 0,1 1,-3 0"
+      android:fillColor="#000000"/>
+  <path
+      android:pathData="M11.08,16.18C10.44,15.9 9.74,15.75 9,15.75s-1.44,0.15 -2.08,0.43C6.36,16.42 6,16.96 6,17.57V18h6v-0.43C12,16.96 11.64,16.42 11.08,16.18z"
+      android:fillColor="#000000"/>
+</vector>
