+++ app/src/main/res/drawable/ic_brapi.xml
@@ -1,0 +1,38 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="28dp"
+    android:height="28dp"
+    android:viewportWidth="60"
+    android:viewportHeight="60">
+  <path
+      android:pathData="m14.3998,-0.0002a14.3998,14.3998 0,0 0,-14.3998 14.3998,14.3998 14.3998,0 0,0 14.3998,14.3998l14.3998,0l0,-14.3998a14.3998,14.3998 0,0 0,-14.3998 -14.3998z"
+      android:strokeLineJoin="miter"
+      android:strokeWidth="0.354"
+      android:fillColor="#006000"
+      android:strokeColor="#00000000"
+      android:fillType="evenOdd"
+      android:strokeLineCap="square"/>
+  <path
+      android:pathData="m45.5994,0.0003a14.3998,14.3998 0,0 1,14.3998 14.3998,14.3998 14.3998,0 0,1 -14.3998,14.3998l-14.3998,0l0,-14.3998a14.3998,14.3998 0,0 1,14.3998 -14.3998z"
+      android:strokeLineJoin="miter"
+      android:strokeWidth="0.354"
+      android:fillColor="#00a040"
+      android:strokeColor="#00000000"
+      android:fillType="evenOdd"
+      android:strokeLineCap="square"/>
+  <path
+      android:pathData="m14.3998,59.9999a14.3998,14.3998 0,0 1,-14.3998 -14.3998,14.3998 14.3998,0 0,1 14.3998,-14.3998l14.3998,0l0,14.3998a14.3998,14.3998 0,0 1,-14.3998 14.3998z"
+      android:strokeLineJoin="miter"
+      android:strokeWidth="0.354"
+      android:fillColor="#905c08"
+      android:strokeColor="#00000000"
+      android:fillType="evenOdd"
+      android:strokeLineCap="square"/>
+  <path
+      android:pathData="m45.5994,59.9994a14.3998,14.3998 0,0 0,14.3998 -14.3998,14.3998 14.3998,0 0,0 -14.3998,-14.3998l-14.3998,0l0,14.3998a14.3998,14.3998 0,0 0,14.3998 14.3998z"
+      android:strokeLineJoin="miter"
+      android:strokeWidth="0.354"
+      android:fillColor="#a0c018"
+      android:strokeColor="#00000000"
+      android:fillType="evenOdd"
+      android:strokeLineCap="square"/>
+</vector>
