+++ app/src/main/res/navigation/navigation.xml
@@ -6,6 +6,8 @@
     android:id="@+id/main_graph"
     app:startDestination="@+id/events_fragment">
 
+    <include app:graph="@navigation/preferences"/>
+
     <!-- Events Fragment is the main entry fragment that shows the entire crossing table -->
     <fragment android:id="@+id/events_fragment"
         android:name="org.phenoapps.intercross.fragments.EventsFragment"
@@ -178,24 +180,6 @@
     </fragment>
 
     <fragment
-        android:id="@+id/settings_fragment"
-        android:name="org.phenoapps.intercross.fragments.SettingsFragment"
-        android:label="@string/settings_label">
-        <action
-            android:id="@+id/action_to_pattern_fragment"
-            app:destination="@id/pattern_fragment"
-            app:launchSingleTop="true" />
-        <action
-            android:id="@+id/action_to_import_zpl_fragment"
-            app:destination="@id/import_zpl_fragment"
-            app:launchSingleTop="true" />
-        <action
-            android:id="@+id/action_to_about"
-            app:destination="@id/about_fragment"
-            app:launchSingleTop="true" />
-    </fragment>
-
-    <fragment
         android:id="@+id/pattern_fragment"
         android:name="org.phenoapps.intercross.fragments.PatternFragment"
         android:label="@string/patterns_label"
@@ -295,9 +279,9 @@
 
     <!-- Global Actions used by Navigation Drawer -->
     <action android:id="@+id/global_action_to_settings_fragment"
-        app:destination="@id/settings_fragment"
+        app:destination="@id/preferences"
         app:launchSingleTop="true"
-        app:popUpTo="@id/settings_fragment"/>
+        app:popUpTo="@id/preferences"/>
 
     <action android:id="@+id/global_action_to_event_detail"
         app:destination="@id/event_fragment"/>
