+++ app/src/main/res/xml/about_preferences.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen
+    android:key="org.phenoapps.intercross.ROOT_PREFERENCES_ABOUT"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto" >
+    <PreferenceCategory android:title="About" app:iconSpaceReserved="false">
+        <Preference
+            android:icon="@drawable/ic_nv_about"
+            android:key="org.phenoapps.intercross.ABOUT"
+            android:title="About" />
+    </PreferenceCategory>
+</PreferenceScreen>
