+++ app/src/main/res/xml/profile_preferences.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen
+    android:key="org.phenoapps.intercross.ROOT_PREFERENCES_PROFILE"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto" >
+    <PreferenceCategory android:title="Profile" app:iconSpaceReserved="false">
+        <EditTextPreference
+            android:title="Person"
+            android:icon="@drawable/ic_setting_person"
+            android:key="org.phenoapps.intercross.PERSON" />
+
+        <EditTextPreference
+            android:icon="@drawable/ic_setting_experiment"
+            android:key="org.phenoapps.intercross.EXPERIMENT"
+            android:title="Experiment" />
+    </PreferenceCategory>
+</PreferenceScreen>
