+++ app/src/main/res/xml/workflow_preferences.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen
+    android:key="org.phenoapps.intercross.ROOT_PREFERENCES_WORKFLOW"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto" >
+    <PreferenceCategory android:title="Workflow" app:iconSpaceReserved="false">
+        <SwitchPreference
+            android:defaultValue="false"
+            android:icon="@drawable/ic_setting_additional_info"
+            android:key="org.phenoapps.intercross.COLLECT_INFO"
+            android:title="Collect additional information"/>
+        <Preference
+            android:icon="@drawable/ic_note_add_black_24dp"
+            android:key="org.phenoapps.intercross.META_DATA"
+            android:title="Create Metadata"/>
+        <Preference
+            android:icon="@drawable/ic_update_black_24dp"
+            android:key="org.phenoapps.intercross.META_DATA_DEFAULTS"
+            android:title="Metadata Defaults"/>
+        <SwitchPreference
+            android:defaultValue="false"
+            android:icon="@drawable/ic_setting_sound"
+            android:key="org.phenoapps.intercross.AUDIO_ENABLED"
+            android:title="Audio notifications"/>
+        <SwitchPreference
+            android:defaultValue="false"
+            android:icon="@drawable/ic_book_open"
+            android:key="org.phenoapps.intercross.OPEN_CROSS_IMMEDIATELY"
+            android:title="Open Cross After Creating"/>
+        <SwitchPreference
+            android:defaultValue="false"
+            android:icon="@drawable/ic_receipt_long_black_24dp"
+            android:key="org.phenoapps.intercross.COMMUTATIVE_CROSSING"
+            android:title="Commutative Crossing"/>
+    </PreferenceCategory>
+</PreferenceScreen>
